Image_models/Density_models/SS-DCNet/requirements.txt
python==3.6.2
pytorch>=0.4.0
numpy==1.14.0
pillow>=6.2.0
scipy==1.0.0
pandas==0.22.0